GLT-4423 Changes requested again in PR.

+++ miso-web/src/main/webapp/scripts/form_library.js
@@ -476,7 +476,7 @@ FormTarget.library = (function ($) {
                     title: "Spike-In Volume",
                     data: "spikeInVolume",
                     type: "decimal",
-                    precision: 16,
+                    precision: 14,
                     scale: 10,
                   },
                 ])
